+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Synchronize scroll with active transcript sentence
+
 ## [3.9.1] - 2020-06-24
 
 ### Fixed